mountd: auth_unix_ip should downcall on error to prevent hangs
Since bf6a4febaa78bf188896b7b5b02c46562dd08b70 "mountd: handle
allocation failures in auth_unix_ip upcall", a failure to map the
address of an incoming client to a name could result in a hang.

We should be responding with an error in the case, not just skipping the
downcall and leaving everybody hanging.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Steve Dickson <steved@redhat.com>