gssd: Provide an introduction in gssd(8)
It's good practice in user documentation to define terms before they
are used.  Add an INTRODUCTION section that defines important terms
that are used in the DESCRIPTION and OPTIONS sections.  The key
concepts are GSS context, user credential, machine credential, and
keytab.

The RFCs I looked at capitalize both "gss" and "rpcsec_gss".  For
consistency I changed this throughout the man page.

Acked-by: J. Bruce Fields <bfields@fieldses.org>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Steve Dickson <steved@redhat.com>