mount.nfs: improve handling of bg, fg, and sloppy in nfsmount.conf
This patch makes 2 small improvements to the parsing of the bg, fg, and
sloppy mount options in nfsmount.conf.

1. "bg" and "fg" negate should each other.  "Background=True" should
mean "bg" and "Background=False" should mean "fg".  The same applies to
"Foreground".

2. Once we see "Sloppy=False" while parsing the configuration file we
should ignore subsequent occurrences of the sloppy option.  This will
preserve the "right-most setting wins" behavior for the sloppy mount
option.

Acked-by: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: Scott Mayhew <smayhew@redhat.com>
Signed-off-by: Steve Dickson <steved@redhat.com>