statd: add options to set port number of lockd
Even though lockd is a totally separate process to statd, they
depended on each other: statd much be running for lockd to be useful.
So an easy way to set the port numbers used by lockd is to get statd
to set them.

This patch add --nlm-port and --nlm-tcp-port to that end.

Signed-off-by: Steve Dickson <steved@redhat.com>