mountd: use SONAME fir libnfsjunct when loading with dlopen
The standard for loading shared libraries is to identify them by their
"soname" (Which "objdump -x $BINARY | grep SONAME" will report).
However mountd currently loads using the "linker name" which should only
be used when building new code.

Future releases of fedfs-utils will define the soname in the include
file, so if that is defined, use it.  If not, use the soname of the
first version: "libnfsjunct.so.0".

This is a slight behavioural change.  However all distros known to
package fedfs-utils will install "libnfsjunct.so.0" whenever they
install the old name of "libnfsjunct.so", and "make install" will
install both.  So it should not be a noticeable change.

Also only test the JP_API_VERSION if it is defined.  As the version is
embedded in the soname, a secondary test is not needed.

Cc: Chuck Lever <chuck.lever@oracle.com>
Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Steve Dickson <steved@redhat.com>